Move some utilities to services
