+++ packages/metrics-app/src/App.tsx
@@ -3,13 +3,13 @@ import { registerRootComponent } from 'expo'
 import { StrictMode, Suspense, useEffect } from 'react'
 import * as SplashScreen from 'expo-splash-screen'
 import { RootSiblingParent } from 'react-native-root-siblings'
+import './services/i18n'
 import { setBackgroundMessageListener, setForegroundMessageListener } from './services/messaging'
 import { performMeasurement } from './services/measurements'
 import Navigator from './navigation/Navigator'
-import './utils/i18n'
 
 
-// Prevent app loading until the stack navigator is ready
+// Keep the splash screen visible until the navigator is ready
 SplashScreen.preventAutoHideAsync()
 
 function App() {
