+++ packages/metrics-app/src/services/tutorial.ts
@@ -1,5 +1,5 @@
 import AsyncStorage from '@react-native-async-storage/async-storage'
-import { logger } from './logger'
+import { logger } from '../utils/logger'
 
 
 // Logger tag
